# nevup-backend/openapi.yaml
openapi: 3.0.3
info:
  title: NevUp Trade Journal API
  description: |
    NevUp Hiring Hackathon 2026 — Mock API Specification.
    
    This spec defines the shared contract between all three tracks.
    Track 3 (Frontend) must consume these endpoints exactly as defined.
    Track 1 (Backend) must implement them exactly as defined.
    Track 2 (AI Engine) uses the behavioral metrics and session data.

    **Running the mock server locally (no account needed):**
    ```
    npx @stoplight/prism-cli mock nevup_openapi.yaml --port 4010
    ```
    All Track 3 endpoints will then be live at http://localhost:4010

  version: "1.0.0"
  contact:
    name: NevUp Hackathon Organizers
servers:
  - url: http://localhost:4010
    description: Local mock server (Prism)
  - url: https://api.nevup.dev/v1
    description: Production (Track 1 live deployment)
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /trades:
    post:
      summary: Submit a new trade
      description: |
        Idempotent on tradeId. Duplicate submissions must return HTTP 200
        with the existing record — not 500 or 409.
      operationId: createTrade
      tags: [Trades]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeInput'
            example:
              tradeId: "550e8400-e29b-41d4-a716-446655440000"
              userId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              sessionId: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
              asset: "AAPL"
              assetClass: "equity"
              direction: "long"
              entryPrice: 178.45
              exitPrice: 182.30
              quantity: 10
              entryAt: "2025-01-06T09:35:00Z"
              exitAt: "2025-01-06T11:20:00Z"
              status: "closed"
              planAdherence: 4
              emotionalState: "calm"
              entryRationale: "Breakout above key resistance with volume confirmation"
      responses:
        '200':
          description: Trade created (or already exists — idempotent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /trades/{tradeId}:
    get:
      summary: Get a single trade by ID
      operationId: getTrade
      tags: [Trades]
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Trade record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── SESSIONS ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued at login. See jwt_format.md for the canonical token structure.
        Every endpoint enforces row-level tenancy — the userId in the JWT must
        match the data being accessed. Cross-tenant reads return HTTP 403.
  schemas:
...
    TradeInput:
      type: object
      required:
        - tradeId
        - userId
        - sessionId
        - asset
        - assetClass
        - direction
        - entryPrice
        - exitPrice
        - quantity
        - entryAt
        - exitAt
        - status
        - planAdherence
        - emotionalState
        - entryRationale
      properties:
        tradeId:
          type: string
          format: uuid
          description: UUIDv4. Submission is idempotent on this field.
        userId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
          description: Groups trades belonging to one trading session.
        asset:
          type: string
          example: "AAPL"
        assetClass:
          type: string
          enum: [equity, crypto, forex]
        direction:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
          format: decimal
          description: decimal(18,8)
        exitPrice:
          type: number
          format: decimal
          nullable: true
          description: null if trade is still open
        quantity:
          type: number
          format: decimal
          description: decimal(18,8)
        entryAt:
          type: string
          format: date-time
          description: ISO-8601 UTC
        exitAt:
          type: string
          format: date-time
          nullable: true
          description: ISO-8601 UTC. null if trade is still open.
        status:
          type: string
          enum: [open, closed, cancelled]
        planAdherence:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
          description: User self-rating at close. null if trade not yet closed.
        emotionalState:
          type: string
          enum: [calm, anxious, greedy, fearful, neutral]
          nullable: true
        entryRationale:
          type: string
          maxLength: 500
          nullable: true
    Trade:
      allOf:
        - $ref: '#/components/schemas/TradeInput'
        - type: object
          properties:
            outcome:
              type: string
              enum: [win, loss]
              nullable: true
            pnl:
              type: number
              nullable: true
              description: Computed P&L. null if trade is open.
            revengeFlag:
              type: boolean
              description: |
                True if trade opened within 90s of a losing close
                AND emotionalState is anxious or fearful.
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    # ── SESSION ────────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [error, message, traceId]
      properties:
        error:
          type: string
          example: "TRADE_NOT_FOUND"
        message:
          type: string
          example: "Trade with the given tradeId does not exist."
        traceId:
          type: string
          description: Structured log correlation ID.
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: |
        Authenticated user does not have access to this resource.
        Cross-tenant reads always return 403 — never 404.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'